fix: getThemeInitScript in separate Server-Datei ausgelagert
"use client"-Modul kann nicht in Server Components importiert werden.
getThemeInitScript() lebt jetzt in src/lib/themeScript.ts (kein "use client").
